Added Abstract class for BeforeEach and AfterEach methods
