+++ src/test/java/io/samsquamptch/afterclass/service/LessonServiceTests.java
@@ -1,23 +1,15 @@
 package io.samsquamptch.afterclass.service;
 
-import io.samsquamptch.afterclass.Group;
 import io.samsquamptch.afterclass.Lesson;
-import io.samsquamptch.afterclass.User;
 import io.samsquamptch.afterclass.WeekDay;
 import io.samsquamptch.afterclass.dto.LessonDTO;
 import io.samsquamptch.afterclass.dto.LessonRequestDTO;
 import io.samsquamptch.afterclass.exception.NotFoundException;
-import io.samsquamptch.afterclass.repositories.GroupRepository;
-import io.samsquamptch.afterclass.repositories.LessonRepository;
-import io.samsquamptch.afterclass.repositories.UserRepository;
 import io.samsquamptch.afterclass.services.LessonService;
 import jakarta.transaction.Transactional;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.jdbc.core.JdbcTemplate;
 
 import java.time.LocalTime;
 import java.util.List;
@@ -27,51 +19,10 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @Transactional
-public class LessonServiceTests {
+public class LessonServiceTests extends AbstractIntegrationTests {
 
     @Autowired
     private LessonService lessonService;
-
-    @Autowired
-    private LessonRepository lessonRepository;
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
-    private GroupRepository groupRepository;
-
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
-
-    @BeforeEach
-    void setUp() {
-        Group group = new Group("Test Group", "zQxprL");
-        Group savedGroup = groupRepository.saveAndFlush(group);
-        User user = new User("Chris");
-        user.setGroup(savedGroup);
-        User savedUser = userRepository.saveAndFlush(user);
-        Lesson lesson1 = new Lesson("Business Law",
-                WeekDay.FRIDAY,
-                LocalTime.of(18,0),
-                LocalTime.of(21,0));
-        Lesson lesson2 = new Lesson("Finance",
-                WeekDay.WEDNESDAY,
-                LocalTime.of(18,0),
-                LocalTime.of(19,30));
-        lesson1.setUser(savedUser);
-        lesson2.setUser(savedUser);
-        lessonRepository.saveAndFlush(lesson1);
-        lessonRepository.saveAndFlush(lesson2);
-    }
-
-    @AfterEach
-    void resetAutoIncrement() {
-        jdbcTemplate.execute("ALTER TABLE lessons ALTER COLUMN id RESTART WITH 1");
-        jdbcTemplate.execute("ALTER TABLE users ALTER COLUMN id RESTART WITH 1");
-        jdbcTemplate.execute("ALTER TABLE groups ALTER COLUMN id RESTART WITH 1");
-    }
-
 
     @Test
     public void GetLesson() {
